+++ src/api-helper/goal/executeGoal.ts
@@ -17,20 +17,20 @@
 import { failure, HandlerContext, HandlerResult, logger, Success } from "@atomist/automation-client";
 import { jwtToken } from "@atomist/automation-client/globals";
 import * as path from "path";
-import { ExecuteGoalResult } from "../../../../api/goal/ExecuteGoalResult";
-import { ExecuteGoalWithLog, RunWithLogContext } from "../../../../api/goal/ExecuteGoalWithLog";
-import { Goal } from "../../../../api/goal/Goal";
-import { SdmGoal } from "../../../../ingesters/sdmGoalIngester";
-import { descriptionFromState, updateGoal } from "../../../../internal/delivery/goals/support/storeGoals";
-import { InterpretLog } from "../../../../spi/log/InterpretedLog";
-import { ProjectLoader } from "../../../../spi/project/ProjectLoader";
-import { spawnAndWatch } from "../../../../util/misc/spawned";
+import { ExecuteGoalResult } from "../../api/goal/ExecuteGoalResult";
+import { ExecuteGoalWithLog, RunWithLogContext } from "../../api/goal/ExecuteGoalWithLog";
+import { Goal } from "../../api/goal/Goal";
+import { SdmGoal } from "../../ingesters/sdmGoalIngester";
+import { InterpretLog } from "../../spi/log/InterpretedLog";
+import { ProjectLoader } from "../../spi/project/ProjectLoader";
+import { spawnAndWatch } from "../../util/misc/spawned";
+import { descriptionFromState, updateGoal } from "./storeGoals";
 
 import { sprintf } from "sprintf-js";
 
 import * as stringify from "json-stringify-safe";
-import { reportGoalError } from "../../../../internal/delivery/goals/support/reportGoalError";
-import { toToken } from "../../../../util/credentials/toToken";
+import { toToken } from "../../util/credentials/toToken";
+import { reportGoalError } from "./reportGoalError";
 
 /**
  * Central function to execute a goal with progress logging
