+++ src/api-helper/goal/storeGoals.ts
@@ -19,13 +19,13 @@ import { RemoteRepoRef } from "@atomist/automation-client/operations/common/Repo
 import { addressEvent } from "@atomist/automation-client/spi/message/MessageClient";
 import * as _ from "lodash";
 import { sprintf } from "sprintf-js";
-import { Goal, hasPreconditions } from "../../../../api/goal/Goal";
-import { disregardApproval } from "../../../../api/goal/support/approvalGate";
-import { GoalImplementation } from "../../../../api/goal/support/SdmGoalImplementationMapper";
+import { Goal, hasPreconditions } from "../../api/goal/Goal";
+import { disregardApproval } from "../../api/goal/support/approvalGate";
+import { GoalImplementation } from "../../api/goal/support/SdmGoalImplementationMapper";
 import {
     GoalRootType, SdmGoal, SdmGoalFulfillment, SdmGoalKey, SdmGoalState,
     SdmProvenance,
-} from "../../../../ingesters/sdmGoalIngester";
+} from "../../ingesters/sdmGoalIngester";
 
 export function environmentFromGoal(goal: Goal) {
     return goal.definition.environment.replace(/\/$/, ""); // remove trailing slash at least
