+++ src/handlers/events/delivery/goals/resetGoals.ts
@@ -20,13 +20,13 @@ import { commandHandlerFrom } from "@atomist/automation-client/onCommand";
 import { GitHubRepoRef } from "@atomist/automation-client/operations/common/GitHubRepoRef";
 import { RemoteRepoRef } from "@atomist/automation-client/operations/common/RepoId";
 import * as stringify from "json-stringify-safe";
+import { chooseAndSetGoals } from "../../../../api-helper/goal/chooseAndSetGoals";
 import { SdmGoalImplementationMapper } from "../../../../api/goal/support/SdmGoalImplementationMapper";
 import { GoalsSetListener } from "../../../../api/listener/GoalsSetListener";
 import { GoalSetter } from "../../../../api/mapping/GoalSetter";
 import { ProjectLoader } from "../../../../spi/project/ProjectLoader";
 import { RepoRefResolver } from "../../../../spi/repo-ref/RepoRefResolver";
 import { PushFields, PushForCommit, RepoBranchTips } from "../../../../typings/types";
-import { chooseAndSetGoals } from "./SetGoalsOnPush";
 
 @Parameters()
 export class ResetGoalsParameters {
@@ -81,7 +81,6 @@ function resetGoalsOnCommit(rules: {
         const id = GitHubRepoRef.from({owner: commandParams.owner, repo: commandParams.repo, sha, branch});
 
         const push = await fetchPushForCommit(ctx, id, commandParams.providerId);
-
         const credentials = {token: commandParams.githubToken};
 
         const goals = await chooseAndSetGoals({
