+++ test/handlers/events/delivery/ExecuteGoalTest.ts
@@ -18,11 +18,11 @@ import { Success } from "@atomist/automation-client";
 import { InMemoryProject } from "@atomist/automation-client/project/mem/InMemoryProject";
 
 import * as assert from "power-assert";
+import { executeGoal } from "../../../../src/api-helper/goal/executeGoal";
 import { createEphemeralProgressLog } from "../../../../src/api-helper/log/EphemeralProgressLog";
 import { RunWithLogContext } from "../../../../src/api/goal/ExecuteGoalWithLog";
 import { Goal } from "../../../../src/api/goal/Goal";
 import { IndependentOfEnvironment } from "../../../../src/api/goal/support/environment";
-import { executeGoal } from "../../../../src/handlers/events/delivery/goals/executeGoal";
 import { SdmGoal } from "../../../../src/ingesters/sdmGoalIngester";
 import { lastLinesLogInterpreter } from "../../../../src/internal/delivery/goals/support/logInterpreters";
 import { fakeContext } from "../../../../src/util/test/fakeContext";
